First pass on file select component
