+++ packages/mux-uploader/src/index.ts
@@ -2,7 +2,8 @@ export * as constants from './constants';
 import MuxUploaderElement, { MuxUploaderElementEventMap } from './mux-uploader';
 import MuxUploaderProgressElement from './mux-uploader-progress';
 import MuxUploaderDropElement from './mux-uploader-drop';
+import MuxUploaderFileSelectElement from './mux-uploader-file-select';
 
-export { MuxUploaderDropElement, MuxUploaderProgressElement };
+export { MuxUploaderDropElement, MuxUploaderProgressElement, MuxUploaderFileSelectElement };
 export type { MuxUploaderElementEventMap };
 export default MuxUploaderElement;
